+++ src/noahtheduke/splint/config.clj
@@ -4,33 +4,40 @@
 
 (ns noahtheduke.splint.config
   (:require
-    [clojure.edn :as edn]
     [clojure.java.io :as io]
     [clojure.string :as str]
+    [edamame.core :as e]
     [noahtheduke.splint.rules :refer [global-rules]])
   (:import
-    (java.io PushbackReader)))
+    (java.io File)))
 
 (set! *warn-on-reflection* true)
 
 (def version (delay (str/trim (slurp (io/resource "SPLINT_VERSION")))))
 (defn splint-version [] (str "splint v" @version))
 
-(defn slurp-1-edn
-  "Read a single value from an edn file. Returns nil if file is empty, throws if file contains more than 1 value."
-  [file]
-  (let [eof (Object.)
-        file-reader (io/reader file)]
-    (with-open [rdr (PushbackReader. file-reader)]
-      (let [v (edn/read {:eof eof} rdr)]
-        (when-not (identical? eof v)
-          (if (identical? eof (edn/read {:eof eof} rdr))
-            v
-            (throw (ex-info "edn file has more than 1 value" {:file file
-                                                              :type :config}))))))))
+(defn slurp-edn
+  "Read values from an edn file.
+
+  Returns nil if file is empty.
+  If :multiple is provided, returns a vector of values.
+  If :multiple is not provided, returns a single value or throws if file contains more than 1 value."
+  ([file] (slurp-edn file nil))
+  ([file {:keys [multiple]}]
+   (let [parse-opts {:read-eval identity
+                     :regex true
+                     :fn true}
+         v (e/parse-string-all (slurp file) parse-opts)]
+     (cond
+       multiple (not-empty v)
+       (< 1 (count v))
+       (throw (ex-info "edn file has more than 1 value" {:file file
+                                                         :type :config}))
+       :else
+       (first v)))))
 
 (defn read-default-config []
-  (slurp-1-edn (io/resource "config/default.edn")))
+  (slurp-edn (io/resource "config/default.edn")))
 
 (def default-config
   (delay (read-default-config)))
@@ -41,7 +48,7 @@
       (if (.exists config)
         {:dir dir
          :file (.getAbsoluteFile config)
-         :local (slurp-1-edn config)}
+         :local (slurp-edn config)}
         (when-let [parent (.getParentFile dir)]
           (recur parent))))))
 
@@ -128,3 +135,83 @@
                       (str " " (str/trim (str/join "\n\n" rule-strs)))
                       "}"])]
     (spit ".splint.edn" new-config)))
+
+(defn parse-clojure-version
+  [version]
+  (let [pat #"(?<major>\d+)\.(?<minor>\d+)\.(?<incremental>\d+)(?:-(?<qualifier>[a-zA-Z0-9_]+))?(?:-(?<snapshot>SNAPSHOT))?"
+        m (re-matcher pat version)
+        _ (.matches m)
+        qualifier (.group m "qualifier")
+        snapshot (if (= qualifier "SNAPSHOT") qualifier (.group m "snapshot"))
+        qualifier (when-not (= qualifier "SNAPSHOT") qualifier)]
+    {:major (parse-long (.group m "major"))
+     :minor (parse-long (.group m "minor"))
+     :incremental (parse-long (.group m "incremental"))
+     :qualifier qualifier
+     :snapshot snapshot}))
+
+(defn project-clojure-version
+  [project-map]
+  (if project-map
+    (when-let [version-str
+               (case (::type project-map)
+                 :deps-edn (some-> (:deps project-map)
+                                   ('org.clojure/clojure)
+                                   :mvn/version)
+                 :project-clj (some->> (:dependencies project-map)
+                                       (filter #(= 'org.clojure/clojure (first %)))
+                                       first
+                                       second)
+                 ; else
+                 nil)]
+      (parse-clojure-version version-str))
+    *clojure-version*))
+
+(defn project-paths
+  "All of the default paths in the project map.
+  For deps.edn, :paths from the base plus :paths or :extra-paths from :dev and :test aliases.
+  For project.clj, :source-paths and :test-paths from base plus both from :dev and :test profiles."
+  [project-map]
+  (when project-map
+    (let [src-paths (case (::type project-map)
+                      :deps-edn (:paths project-map)
+                      :project-clj (into (or (:source-paths project-map)
+                                             ["src"])
+                                         (or (:test-paths project-map)
+                                             ["test"]))
+                      ; else
+                      nil)
+          other-paths (case (::type project-map)
+                        :deps-edn (let [aliases (:aliases project-map)]
+                                    (into (some-> aliases :dev :extra-paths)
+                                          (some-> aliases :test :extra-paths)))
+                        :project-clj
+                        (let [profiles (:profiles project-map)]
+                          (concat
+                            (when (map? (:dev profiles))
+                              (into (some-> profiles :dev :source-paths)
+                                    (some-> profiles :dev :test-paths)))
+                            (when (map? (:test profiles))
+                              (into (some-> profiles :test :source-paths)
+                                    (some-> profiles :test :test-paths)))))
+                        ; else
+                        nil)]
+      (into (vec src-paths) other-paths))))
+
+(defn read-project-file
+  "Read the first available proejct config file.
+  In order, checks deps.edn, project.clj, boot."
+  [^File deps-edn ^File project-clj]
+  (let [project-file
+        (cond
+          (and deps-edn (.exists deps-edn))
+          (assoc (slurp-edn deps-edn) ::type :deps-edn)
+          (and project-clj (.exists project-clj))
+          (let [v (->> (slurp-edn project-clj {:multiple true})
+                       (filter #(and (seq? %) (= 'defproject (first %))))
+                       first
+                       (drop 3)
+                       (apply hash-map))]
+            (assoc v ::type :project-clj)))]
+    {:clojure-version (project-clojure-version project-file)
+     :paths (project-paths project-file)}))