+++ src/noahtheduke/splint/rules/style/prefer_clj_math.clj
@@ -91,6 +91,7 @@
   {:pattern '%math?%-?sym
    :init-type :symbol
    :message "Use the `clojure.math` function instead of interop."
+   :min-clojure-version {:major 1 :minor 11}
    :on-match (fn [ctx rule form {:syms [?sym]}]
                (let [new-form (Math->clj-math ?sym)]
                  (->diagnostic ctx rule form {:replace-form new-form})))})