+++ test/noahtheduke/splint/re_frame_test.clj
@@ -33,7 +33,10 @@
 (defexpect ^:integration re-frame-test
   (let [re-frame (gl/procure "https://github.com/day8/re-frame.git" 'day8/re-frame "v1.3.0")
         results (run-impl [re-frame]
-                          (assoc all-enabled-config :silent true :parallel false))]
+                          (assoc all-enabled-config
+                                 :silent true
+                                 :parallel false
+                                 :clojure-version *clojure-version*))]
     (expect
       (match?
         (m/equals re-frame-diagnostics)